Sépare l'affichage des signes vitaux dans leur propre panel
